criacao dos modulos para desacoplar o codigo
